backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
groq
pydantic
httpx
